+++ SPO/UpperLevel/GUI/Frames/widgets.h
@@ -4,11 +4,11 @@
 #include "main.h"
 #include "layout.h"
 
-button_t drawFillButton (uint16_t xPos, uint16_t yPos, uint16_t xSize, uint16_t ySize, uint16_t* label, bool isTouch);
+button_t drawFillButton (uint16_t xPos, uint16_t yPos, uint16_t xSize, uint16_t ySize, uint8_t* label, bool isTouch);
 
-button_t drawTextLabel (uint16_t xPos, uint16_t yPos, uint16_t xSize, uint16_t ySize, uint16_t* label);
+button_t drawTextLabel (uint16_t xPos, uint16_t yPos, uint16_t xSize, uint16_t ySize, uint8_t* label);
 
-void drawMainBar(bool returnBut, uint16_t xPosLogo, uint16_t yPosLogo, uint16_t* label);
+void drawMainBar(bool returnBut, uint16_t xPosLogo, uint16_t yPosLogo, uint8_t* label);
 
 void drawMainWindow();
 
@@ -20,7 +20,7 @@ typedef enum
 }statusBar_t;
 void drawStatusBarOkCancel();
 
-void drawStatusBarSave(uint16_t* label);
+void drawStatusBarSave(uint8_t* label);
 
 void drawStatusBarEmpty();
